e2e reorganization
