+++ e2e/__tests__/topbar.spec.ts
@@ -1,7 +1,7 @@
 import {expect, test} from '@playwright/test'
 
-import {logoSelector, logoUrl} from './constants'
-import {waitMicrolcLoaded} from './utils'
+import {logoSelector, logoUrl} from '../utils/constants'
+import {waitMicrolcLoaded} from '../utils/utils'
 
 test('Logo is loaded correctly', async ({page}) => {
   await waitMicrolcLoaded(page)
