+++ e2e/__tests__/usermenu.spec.ts
@@ -1,7 +1,7 @@
 import {expect, test} from '@playwright/test'
 
-import {userMenuSelector} from './constants'
-import {waitMicrolcLoaded} from './utils'
+import {userMenuSelector} from '../utils/constants'
+import {waitMicrolcLoaded} from '../utils/utils'
 
 test('User menu is not in page', async ({page}) => {
   await waitMicrolcLoaded(page)
