+++ e2e/config.ts
@@ -1,7 +1,8 @@
 import {ChromiumEnv, FirefoxEnv, setConfig, test, WebKitEnv} from '@playwright/test'
+import path from 'path'
 
 setConfig({
-  testDir: __dirname,
+  testDir: path.join(__dirname, '__tests__'),
   timeout: 10000,
 })
 
@@ -12,4 +13,4 @@ const options = {
 
 test.runWith(new ChromiumEnv(options), {tag: 'chromium'});
 test.runWith(new FirefoxEnv(options), {tag: 'firefox'});
-// test.runWith(new WebKitEnv(options), {tag: 'webkit'});
+test.runWith(new WebKitEnv(options), {tag: 'webkit'});
